app/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
pydub==0.25.1
requests==2.32.5
openai==2.16.0
gradio_client==2.0.3
soundfile==0.13.1
numpy==2.2.6
librosa==0.11.0
transformers==4.57.3
peft==0.18.1
faster-whisper==1.2.1
python-multipart==0.0.22
aiofiles==24.1.0
python-docx==1.1.2
pytest==8.4.1
